arreglar la parte de usuarios

+++ InnovatecWebFrontend/src/app/components/panel-de-control/crud-proyectos/proyecto/proyecto.component.ts
@@ -11,6 +11,8 @@ import { Resultado } from 'src/app/shared/models/resultado';
 import { finalize } from 'rxjs/operators';
 import { Grafica } from 'src/app/shared/models/grafica';
 import { PreResultado } from 'src/app/shared/models/pre-resultado';
+declare const $: any;
+declare const M;
 
 @Component({
   selector: 'app-proyecto',
@@ -18,10 +20,10 @@ import { PreResultado } from 'src/app/shared/models/pre-resultado';
   styleUrls: ['./proyecto.component.css']
 })
 export class ProyectoComponent implements OnInit {
+  enableSubmit = true;
   linksDeDescarga: string[];
   filelist = [];
-  fileResultado: any;
-  filelistResultado: PreResultado[];
+  fileImgResultado: any;
   pathfilelist: string[] = [];
   uploads: any[];
   pathfilelistResultados: string[] = [];
@@ -30,12 +32,12 @@ export class ProyectoComponent implements OnInit {
   formProyecto: NgForm;
   formFile: FormGroup;
   resultadosList: Resultado[] = [];
-  preResultadosList: Resultado[] = [];
-  autores: string[];
-  objetivosE: string[];
-  enableSubmit = true;
-  infoGraficas: Grafica[];
-  infoResultado: string;
+
+  elemsAutores: any;
+  instancesAutores: any;
+  elemsObjE: any;
+  instancesObjE: any;
+
   constructor(public fb: FormBuilder, public proyectoService: ProyectoService,  private snackBar: MatSnackBar,
               public storage: AngularFireStorage) {
     this.formFile = this.fb.group({
@@ -44,46 +46,90 @@ export class ProyectoComponent implements OnInit {
   }
 
   ngOnInit() {
+    this.proyectoService.aEproyecto = 'Añadir Proyecto';
     this.proyectoService.getProyectos();
     this.resetForm();
+    this.initializeChips();
     AngularFireModule.initializeApp(environment.firebase);
   }
 
-  resetForm(proyectoForm?: NgForm) {
+
+
+  public resetForm(proyectoForm?: NgForm) {
     if (proyectoForm != null) {
       // tslint:disable-next-line: no-unused-expression
       proyectoForm.reset;
       this.proyectoService.selectProyecto = new Proyecto();
+
+      const chipsAutores = M.Chips.getInstance($('.chips-autores')).chipsData;
+      const chipsObjE = M.Chips.getInstance($('.chips-objetivos-especificos')).chipsData;
+
+      const chipsAutoresI = M.Chips.getInstance($('.chips-autores'));
+      const chipsObjEI = M.Chips.getInstance($('.chips-objetivos-especificos'));
+
+      const index1 = chipsAutores.length;
+      const index2 = chipsObjE.length;
+
+      for (let a = 0; a < 5; a++) {
+        for (let i = 0; i < index1; i++) {
+          chipsAutoresI.deleteChip(i);
+
+        }
+
+        for (let i = 0; i < index2; i++) {
+          chipsObjEI.deleteChip(i);
+
+        }
+      }
+
+      this.proyectoService.aEproyecto = 'Añadir Proyecto';
     }
   }
+
 
   importImages(event) {
     // reset the array
     this.filelist = event.target.files;
   }
 
-  importImageResultado(event) {
-    this.filelistResultado = event.target.files[0];
-  }
-
   onSubmit(formProyecto: NgForm) {
-    this.formProyecto = formProyecto;
-    this.uploads = [];
-    this.pathfilelist = [];
-    const allPercentage: Observable<number>[] = [];
-    // tslint:disable-next-line: prefer-for-of
-    if (this.formProyecto.value.skey != null ) {
-      const tmpList: string[] = Object.values(this.proyectoService.selectProyecto.pathImgsProyecto);
-      const tmpListResultados: Resultado[] = Object.values(this.proyectoService.selectProyecto.resultados);
-      tmpList.forEach(element => {
-        this.proyectoService.deleteFileStorage(element);
+    const chipsAutores = M.Chips.getInstance($('.chips-autores')).chipsData;
+    const autores: string[] = [];
+    chipsAutores.forEach(element1 => {
+      autores.push(element1.tag);
+    });
+
+    if (autores.length <= 0) {
+      this.snackBar.open('Error.', 'Por favor ingrese al menos un(a) autor(a).' , {
+        duration: 2000,
+        panelClass: ['red-snackbar']
       });
-      tmpListResultados.forEach(element => {
-        this.proyectoService.deleteFileStorage(element.pathImgResultado);
-      });
+    } else {
+
+      if (this.proyectoService.preResultadosList.length <= 0 ) {
+        this.snackBar.open('Error.', 'Por favor ingrese al menos un resultado.' , {
+          duration: 2000,
+          panelClass: ['red-snackbar']
+        });
+
+      } else {
+        this.formProyecto = formProyecto;
+        this.uploads = [];
+        this.pathfilelist = [];
+        if (this.formProyecto.value.skey != null ) {
+          const tmpList: string[] = Object.values(this.proyectoService.selectProyecto.pathImgsProyecto);
+          const tmpListResultados: Resultado[] = Object.values(this.proyectoService.selectProyecto.resultados);
+          tmpList.forEach(element => {
+            this.proyectoService.deleteFileStorage(element);
+          });
+          tmpListResultados.forEach(element => {
+            this.proyectoService.deleteFileStorage(element.pathImgResultado);
+          });
+        }
+        this.insertImgs();
+        this.enableSubmit = false;
+      }
     }
-    this.insertImgs();
-    this.enableSubmit = false;
 
   }
 
@@ -93,13 +139,7 @@ export class ProyectoComponent implements OnInit {
       const path = 'proyectos/' + `${file.name}` + ' ' + dateNow;
       const ref = this.storage.ref(path);
       const task = this.storage.upload(path, file);
-      // tslint:disable-next-line: variable-name
-      const _percentage$ = task.percentageChanges();
       this.pathfilelist.push(path);
-      const uploadTrack = {
-        fileName: file.name,
-        percentage: _percentage$
-      };
       task.snapshotChanges().pipe( finalize(() => this.obtenerURL(ref.getDownloadURL()) )).subscribe();
     }
 
@@ -116,7 +156,7 @@ export class ProyectoComponent implements OnInit {
   }
 
   insertImgsResultados() {
-    this.filelistResultado.forEach(element => {
+    this.proyectoService.preResultadosList.forEach(element => {
       this.resultadosList = [];
       const dateNow = Date.now();
       const path = 'proyectos/resultados/' + `${element.fileImgResultado.name}` + ' ' + dateNow;
@@ -128,31 +168,69 @@ export class ProyectoComponent implements OnInit {
 
   }
 
+  initializeChips() {
+
+    $('.chips-autores').chips({
+      placeholder: ' ',
+      secondaryPlaceholder: '+',
+    });
+    $('.chips-objetivos-especificos').chips({
+      placeholder: ' ',
+      secondaryPlaceholder: '+',
+    });
+
+    document.addEventListener('DOMContentLoaded', () => {
+      this.elemsAutores = document.querySelectorAll('.chips-autores');
+      this.instancesAutores = M.Chips.init(this.elemsAutores, Option);
+    });
+
+    document.addEventListener('DOMContentLoaded', () => {
+      this.elemsObjE = document.querySelectorAll('.chips-objetivos-especificos');
+      this.instancesObjE = M.Chips.init(this.elemsObjE, Option);
+    });
+
+  }
+
+
+
+  insertGrafica(formGrafica: NgForm) {
+    const grafica = new Grafica();
+    grafica.linkDeLaGrafica = formGrafica.value.linkDeLaGrafica;
+    grafica.tipoDeGrafica = formGrafica.value.tipoDeGrafica;
+    this.proyectoService.infoGraficas.push(grafica);
+  }
+
   obtenerURLImgResultado(url: Observable<string | null>, path: string ) {
     url.forEach(element => {
       const resultado = new Resultado();
       const index = this.resultadosList.length;
       resultado.imgResultado = element;
       resultado.pathImgResultado = path;
-      resultado.infoResultado = this.preResultadosList[index].imgResultado;
+      resultado.infoResultado = this.proyectoService.preResultadosList[index].infoResultado;
       this.resultadosList.push(resultado);
-      if (this.resultadosList.length === this.filelistResultado.length) {
+      if (this.resultadosList.length === this.proyectoService.preResultadosList.length) {
         this.insertInfo();
       }
     });
   }
 
-  insertPreResultado() {
-    const resultado = new Resultado();
-    resultado.imgResultado = this.fileResultado;
-    this.fileResultado = null;
-    resultado.infoResultado = this.infoResultado;
-    this.infoResultado = '';
-    this.preResultadosList.push(resultado);
-  }
 
   insertInfo() {
     this.linksDeDescarga = [];
+
+    const chipsAutores = M.Chips.getInstance($('.chips-autores')).chipsData;
+    const autores: string[] = [];
+    chipsAutores.forEach(element1 => {
+      autores.push(element1.tag);
+    });
+
+    const chipsObjE = M.Chips.getInstance($('.chips-objetivos-especificos')).chipsData;
+    const objetivosE: string[] = [];
+    chipsObjE.forEach(element1 => {
+      objetivosE.push(element1.tag);
+    });
+
+
     for (let i = 0; i < this.uploads.length; i++) {
       const element = this.uploads[i];
       this.linksDeDescarga.push(element);
@@ -161,12 +239,14 @@ export class ProyectoComponent implements OnInit {
         proyecto.titulo = this.formProyecto.value.titulo;
         proyecto.imgsProyecto = this.uploads;
         proyecto.pathImgsProyecto = this.pathfilelist;
-        proyecto.autores = this.autores;
+
+        proyecto.autores = autores;
         proyecto.resumen = this.formProyecto.value.resumen;
         proyecto.objetivoG = this.formProyecto.value.objetivoG;
-        proyecto.objetivosE = this.objetivosE;
+        proyecto.objetivosE = objetivosE;
+        proyecto.resumenResultados = this.formProyecto.value.resumenResultados;
         proyecto.resultados = this.resultadosList;
-        proyecto.infoGraficas = this.infoGraficas;
+        proyecto.infoGraficas = this.proyectoService.infoGraficas;
 
         if (this.formProyecto.value.skey == null) {
           this.proyectoService.insertProyecto(proyecto);
@@ -177,11 +257,13 @@ export class ProyectoComponent implements OnInit {
         this.resetForm(this.formProyecto);
         this.formFile.reset();
         this.enableSubmit = true;
-        this.snackBar.open('Operación exitosa.', 'Se ha añadido una nueva novedad.' , {
+        this.proyectoService.aEproyecto = 'Añadir Proyecto';
+        this.snackBar.open('Operación exitosa.', 'Se ha añadido un nuevo proyecto.' , {
           duration: 2000,
           panelClass: ['green-snackbar']
         });
       }
     }
+
   }
 }
